+++ DATA/BRONZE/StockData/TATAPOWER_2024_08.csv
@@ -16,4 +16,4 @@ date,open,high,low,close,volume,dividends,stock_splits
 2024-08-22 00:00:00+05:30,426.0,426.05,421.9,422.85,199822,0.0,0.0
 2024-08-23 00:00:00+05:30,423.0,423.0,417.4,418.0,228889,0.0,0.0
 2024-08-26 00:00:00+05:30,420.1,423.45,417.0,422.45,760826,0.0,0.0
-2024-08-27 00:00:00+05:30,422.45,429.6,421.1,428.0,1179392,0.0,0.0
+2024-08-27 00:00:00+05:30,422.45,429.6,421.1,428.0,1179449,0.0,0.0
