+++ DATA/SILVER/StockPrice/StockPrice_data.csv
@@ -3605,7 +3605,7 @@ date,symbol,open,high,low,close,volume
 2024-08-22,TATAPOWER,426.0,426.05,421.9,422.85,199822
 2024-08-23,TATAPOWER,423.0,423.0,417.4,418.0,228889
 2024-08-26,TATAPOWER,420.1,423.45,417.0,422.45,760826
-2024-08-27,TATAPOWER,422.45,429.6,421.1,428.0,1179392
+2024-08-27,TATAPOWER,422.45,429.6,421.1,428.0,1179449
 2021-06-01,VOLTAS,990.82,1000.13,986.7,993.61,619320
 2021-06-02,VOLTAS,991.31,1018.81,988.76,1014.4,1335354
 2021-06-03,VOLTAS,1021.85,1097.02,1016.9,1090.06,7137481
